requirements.txt
asgiref==3.3.4
boto3==1.17.71
botocore==1.20.71
Django==3.2
django-storages==1.11.1
djangorestframework==3.12.4
jmespath==0.10.0
mysqlclient==2.0.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
djangorestframework-simplejwt==4.6.0
django_filter==2.4.0